requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
langchain-ollama>=0.0.1
langchain-chroma>=0.0.1
python-dotenv>=1.0.0
streamlit>=1.31.0
chromadb>=0.4.22
tiktoken>=0.5.2
unstructured>=0.10.30
python-magic>=0.4.27
deepseek-ai==0.0.1
pdfminer.six>=20250506
torch>=2.2.0 